Enable the installer class to write a file in a symbolic directory
The file installer class uses the "mkdirp" function from the "util"
module when vivifying the destination directory of a file to install.
But "mkdirp" chokes on symbolic directory links with a "file exists"
error. This is a potential source of problems: for instance, it prevents
the file installer class from writing a file into the "/etc" directory.
The fix is simple: just use "fs.realpath" instead of "path.normalize".