Fix bad content body (EOFError) when uploading a file.
Should fix issue #125.

The main issue is duplicated data chunks in the resulting body which gets passed to Nack. The bug can be observed mainly for larger files and depends on async IO events and that's why it is not 100% predictable. The duplication occurs due to redundant re-emitting of the 'data' event by the custom-written util.pause(stream) function.

The custom util.pause()/resume() are needed at least for the tests to pass. When removed from RackApplication#handle, Pow seems to work correctly and does not manifest the bug in question, but the tests within test_rack_application.coffee hang and just timeout. I suppose this is due to the 'end' event not being paused by Stream#pause().