Grab the port from Foreman instead of test app (oops!)
The side effect of this change is that I'll know the port number before
the app may have started to listen on it. To fix that, I introduced a
'spawning' state after I know the ports, but before I've confirmed
they're working. I switch to 'ready' once all the ports are up.

This requires changes to Foreman, though. :-(  Watch ddollar/foreman#70.