Fix reinstall/upgrade on OS X Yosemite
Fixes #473

When running the install script multiple times on Yosemite (10.10), the
`launchctl bootstrap` commands fail with a 133 exit code and a message
“Operation already in progress”.

I’ve found that the `bootstrap`/`enable` commands can be replaced with
`load -Fw` to ensure the job is running.

`kickstart -k` is still needed to restart the job in the case of an
upgrade or re-installation.